refactor: Moving address creation to own controller
